+++ analysis/scripts/summarize_autism_dd.R
@@ -71,7 +71,7 @@ posterior_threshold_summary <- function(model, genetic_data) {
   )
 }
 bootstrap_model <- function(model, genetic_data, iteration) {
-  sample_indices <- model$bootstrap_output$bootstrap_samples[, iteration]
+  sample_indices <- model$bootstrap_output$bootstrap_indices[, iteration]
   model_boot <- model
   model_boot$conditional_likelihood <- model$conditional_likelihood[sample_indices, , drop = FALSE]
   model_boot$features <- model$features[sample_indices, , drop = FALSE]